request: Rename files to follow the prefix-name rule
Renames the request files to xdp-request.
