+++ src/account.c
@@ -34,7 +34,7 @@
 #include <gio/gio.h>
 
 #include "account.h"
-#include "request.h"
+#include "xdp-request.h"
 #include "documents.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
