+++ src/background.c
@@ -32,7 +32,7 @@
 #include "xdp-background-monitor.h"
 #include "flatpak-instance.h"
 #include "permissions.h"
-#include "request.h"
+#include "xdp-request.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
 #include "xdp-app-info.h"
