+++ src/email.c
@@ -35,7 +35,7 @@
 #include <gio/gunixfdlist.h>
 
 #include "email.h"
-#include "request.h"
+#include "xdp-request.h"
 #include "documents.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
