+++ src/global-shortcuts.c
@@ -26,7 +26,7 @@
 #include <glib-object.h>
 
 #include "global-shortcuts.h"
-#include "request.h"
+#include "xdp-request.h"
 #include "session.h"
 #include "permissions.h"
 #include "xdp-dbus.h"
