+++ src/location.c
@@ -30,7 +30,7 @@
 #include <gio/gdesktopappinfo.h>
 
 #include "location.h"
-#include "request.h"
+#include "xdp-request.h"
 #include "permissions.h"
 #include "xdp-dbus.h"
 #include "xdp-utils.h"
