+++ src/open-uri.c
@@ -38,7 +38,7 @@
 #include <gio/gdesktopappinfo.h>
 
 #include "open-uri.h"
-#include "request.h"
+#include "xdp-request.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
 #include "xdp-utils.h"
