+++ src/screenshot.c
@@ -36,7 +36,7 @@
 
 #include "screenshot.h"
 #include "permissions.h"
-#include "request.h"
+#include "xdp-request.h"
 #include "documents.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
