+++ src/secret.c
@@ -35,7 +35,7 @@
 #include <gio/gunixfdlist.h>
 
 #include "secret.h"
-#include "request.h"
+#include "xdp-request.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
 #include "xdp-utils.h"
