+++ src/session.h
@@ -20,7 +20,7 @@
 
 #pragma once
 
-#include "request.h"
+#include "xdp-request.h"
 #include "xdp-call.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
