+++ tests/backend/print.c
@@ -7,7 +7,7 @@
 
 #include "xdp-impl-dbus.h"
 
-#include "request.h"
+#include "xdp-request.h"
 #include "print.h"
 
 typedef struct {
