+++ tests/backend/screenshot.c
@@ -7,7 +7,7 @@
 
 #include "xdp-impl-dbus.h"
 
-#include "request.h"
+#include "xdp-request.h"
 #include "screenshot.h"
 
 typedef struct {
