+++ tests/backend/xdp-request.c
@@ -21,7 +21,7 @@
  *       Matthias Clasen <mclasen@redhat.com>
  */
 
-#include "request.h"
+#include "xdp-request.h"
 
 #include <string.h>
