+++ tests/meson.build
@@ -63,7 +63,6 @@ test(
 test_backends = executable(
   'test-backends',
   'backend/test-backends.c',
-  'backend/request.c',
   'backend/session.c',
   'backend/access.c',
   'backend/account.c',
@@ -78,6 +77,7 @@ test_backends = executable(
   'backend/screenshot.c',
   'backend/settings.c',
   'backend/wallpaper.c',
+  'backend/xdp-request.c',
   'glib-backports.c',
   document_portal_built_sources,
   document_portal_built_sources,
